+++ src/tcerror.h
@@ -250,6 +250,8 @@ DEFINE_TCError(SubCmdAttrType, "`SubCmd' attribute is only given to CLI record t
 DEFINE_TCError(EmptyRedirArgs, "not allow arguments after empty command redirection");
 DEFINE_TCError(NoEmptyRedir, "empty command redirection is only allowed after '$(', '<(', '>('");
 DEFINE_TCError(InvalidEnvName, "env name must be valid identifier: `%s'");
+DEFINE_TCError(InvalidMutualGroupElement,
+               "mutual recursive group only allow named function or user-defined command");
 
 DEFINE_TCWarn(MeaninglessCast, "meaningless cast op");
 DEFINE_TCWarn(VarShadowing, "`%s' hides already defined name of outer scope");