+++ src/type_checker.h
@@ -788,7 +788,7 @@ private:
   void visitPrefixAssignNode(PrefixAssignNode &node) override;
   void visitFunctionNode(FunctionNode &node) override;
   void visitUserDefinedCmdNode(UserDefinedCmdNode &node) override;
-  void visitFuncListNode(FuncListNode &node) override;
+  void visitMutualGroupNode(MutualGroupNode &node) override;
   void visitSourceNode(SourceNode &node) override;
   void visitSourceListNode(SourceListNode &node) override;
   void visitCodeCompNode(CodeCompNode &node) override;